+++ database/migrations/2016_01_23_232655_create_checkpoints_table.php
@@ -25,8 +25,8 @@ class CreateCheckpointsTable extends Migration
             $table->string('artist');
             $table->string('description');
             $table->string('image_src');
-            $table->decimal('lat', 20, 20);
-            $table->decimal('lon', 20, 20);
+            $table->double('lat', 20, 20);
+            $table->double('lon', 20, 20);
             $table->string('qr');
             $table->integer('status');
             $table->timestamps();
